+++ odoo/local-src/stock_valuation/wizards/stock_valuation_history.py
@@ -218,63 +218,11 @@ class StockHistory(models.Model):
     )
 
     product_last_in_date = fields.Datetime(
-        'Last Purchasing Date', compute='_get_product_last_in_date'
+        'Last Purchasing Date', related='product_id.product_last_in_date'
     )
     product_last_out_date = fields.Datetime(
-        'Last Selling Date', compute='_get_product_last_out_date'
+        'Last Selling Date', related='product_id.product_last_out_date'
     )
-
-    def _get_product_last_in_date(self):
-        if 'history_date' not in self._context:
-            return
-        self._cr.execute(
-            """
-            SELECT DISTINCT ON (product_id)
-                purchase_order_line.product_id,
-                purchase_order.date_order
-            FROM purchase_order_line
-            LEFT JOIN purchase_order
-                ON purchase_order_line.order_id=purchase_order.id
-            WHERE price_unit > 0
-              AND purchase_order.date_order <= %s
-              AND purchase_order_line.state in ('purchase', 'done')
-            ORDER BY
-                purchase_order_line.product_id,
-                purchase_order.date_order desc
-            """,
-            (self._context['history_date'],),
-        )
-        dates_by_product = dict(self._cr.fetchall())
-        for rec in self:
-            rec.product_last_in_date = dates_by_product.get(
-                rec.product_id.id, False
-            )
-
-    def _get_product_last_out_date(self):
-        if 'history_date' not in self._context:
-            return
-        self._cr.execute(
-            """
-            SELECT DISTINCT ON (product_id)
-                sale_order_line.product_id,
-                sale_order.confirmation_date
-            FROM sale_order_line
-            LEFT JOIN sale_order
-                ON sale_order_line.order_id=sale_order.id
-            WHERE price_unit > 0
-              AND confirmation_date <= %s
-              AND sale_order_line.state != 'cancel'
-            ORDER BY
-                sale_order_line.product_id,
-                sale_order.confirmation_date desc
-            """,
-            (self._context['history_date'],),
-        )
-        dates_by_product = dict(self._cr.fetchall())
-        for rec in self:
-            rec.product_last_out_date = dates_by_product.get(
-                rec.product_id.id, False
-            )
 
     def init(self):
         tools.drop_view_if_exists(self.env.cr, self._table)